feat: add term validity column and refactor component for readability
Added a new column for term validity in the table.
Moved the component from `index.ts` to `page.ts` to improve file readability.
